Refactor playlists room manager to handle dummy API requests
